Fixed hugepage related sva mmu panics + code cleanup
- Turns out the (PMD) hugepage is either explicitly declared as an L1 page or gets internally
declared as an L1 by the nested kernel due to its mappings being set using a set_pte instead of
a set_pmd. Hence, while splitting the large page, we have to remove it from the nested kernel metadata
and redeclare it as an L2
- The newly allocated page, which is used as the L1 page after the hugepage split. The stale values
in this page, cause a problem when the nested kernel tries to update a mapping in this page and updates
the "previously mapped page" by indexing into an invalid pfn in the page descriptor. Fix: Set the page's contents to 0's